Use Puppet::Resource instead of Puppet::Type to find all users
This changes which API is used to find all user accounts. The
Puppet::Type.type() call was a little funky to work with and seemed to
sit a rather low level of abstraction.

The Puppet::Resource.indirection.search() call is used by the 'puppet
resource' command and seems to be at a higher level of abstraction. It
also has outputs that are easier to parse.

https://github.com/puppetlabs/puppet/blob/6.4.2/lib/puppet/application/resource.rb#L228

Additionally, this change gets rid of the improper use of `find_all()`.
We now search through the list of users with `find()` which has the
benefit of stopping the search once a match is found.